Add code for our grabber and flippers in driving code
Works fine BUT the arm is now too heavy to be able to lift
stones with servos, so we are working on redesigning it with
a real motor, and I will have to modify the code for that
to work. Also, the current initialization positions are a bit
weird but that can be fixed easily.